SMP-916: Separar los módulos del plugin en diferentes archivos
- Moved CordovaPluginSentinelSDK file into another folder

- Create AuthModule

- Changed the target-dir where the kt files will locate

- Changed how the packagename is referred
