+++ plugin.xml
@@ -13,10 +13,11 @@
             <preference name="GradlePluginKotlinVersion" value="1.9.10" />
 
             <feature name="CordovaPluginSentinelSDK">
-                <param name="android-package" value="cordova.plugin.sentinel.sdk.CordovaPluginSentinelSDK" />
+                <param name="android-package" value="org.apache.cordova.plugin.CordovaPluginSentinelSDK" />
             </feature>
         </config-file>
         <config-file parent="/*" target="AndroidManifest.xml" />
-        <source-file src="src/android/CordovaPluginSentinelSDK.kt" target-dir="src/main/kotlin/CordovaPluginSentinelSDK" />
+        <source-file src="src/android/org/apache/cordova/plugin/sentinel/CordovaPluginSentinelSDK.kt" target-dir="java/org/apache/cordova/plugin/sentinel/" />
+        <source-file src="src/android/org/apache/cordova/plugin/sentinel/modules/AuthModule.kt" target-dir="java/org/apache/cordova/plugin/sentinel/" />
     </platform>
 </plugin>
