+++ src/android/org/apache/cordova/plugin/sentinel/modules/AuthModule.kt
@@ -1,16 +1,13 @@
-package cordova.plugin.sentinel.sdk
+package org.apache.cordova.plugin.modules
 
-import android.app.Application
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.launch
-import kotlinx.serialization.json.Json
 import org.apache.cordova.CallbackContext
 import org.apache.cordova.CordovaPlugin
 import org.json.JSONArray
 import org.koin.core.context.GlobalContext.startKoin
-import org.koin.dsl.module
 
 import com.qxdev.sentinel_sdk.cloud.data.auth.User
 import com.qxdev.sentinel_sdk.cloud.data.auth.UserData
@@ -18,8 +15,7 @@ import com.qxdev.sentinel_sdk.cloud.data.Response
 import com.qxdev.sentinel_sdk.di.Koin
 import com.qxdev.sentinel_sdk.SentinelProvider
 
-
-class AuthModule{
+class AuthModule {
    private val coroutineScope = CoroutineScope(Dispatchers.IO + Job())
 
    private val cloudProvider = SentinelProvider.cloudProvider
@@ -46,7 +42,7 @@ class AuthModule{
             if(result.success){
                callbackContext.success(result.data.toString())
             } else {
-               callbackContext.error(result.errorCode.toString())
+               callbackContext.error("Hola Ari")
             }
          }else{
             callbackContext.error("Request failed with Exception ${responseResult.exceptionOrNull()?.message}")
@@ -110,42 +106,4 @@ class AuthModule{
          }
       }
    }
-}
-
-class CordovaPluginSentinelSDK : CordovaPlugin() {
-
-   private val authModule = AuthModule()
-
-
-   override fun execute(
-         action: String,
-         args: JSONArray,
-         callbackContext: CallbackContext
-   ): Boolean {
-      if (action == "coolMethod") {
-         val firstNumber = args.getInt(0)
-         val secondNumber = args.getInt(1)
-         this.coolMethod(firstNumber, secondNumber, callbackContext)
-         return true
-      }
-      if (action == "signIn") {
-         val username = args.getString(0)
-         val password = args.getString(1)
-         authModule.signIn(username, password, callbackContext)
-         return true
-      }
-      return false
-   }
-
-   private fun coolMethod(firstNumber: Int, secondNumber: Int, callbackContext: CallbackContext) {
-      if (firstNumber !== null && secondNumber !== null) {
-         val result = firstNumber + secondNumber
-         val message = "El resultado de $firstNumber + $secondNumber es $result"
-         callbackContext.success(message)
-      } else {
-         callbackContext.error("Expected one non-empty string argument.")
-      }
-   }
-
-
 }
